Revert "lister voiture Acceuil"
This reverts commit 27653ad758c0c0761728cd8cdd482c4580966735.

+++ serveur/inventaireVoiture/includes/Voiture.inc.php
@@ -9,7 +9,7 @@ class Voiture {
     private $prix;
     private $quantite;
 
-    function __construct(int $idV, string $nomVoiture, string $description, string $image, float $prix, int $quantite) {
+    function __construct(int $idV, string $nomVoiture, string $description, string $image, double $prix, int $quantite) {
         $this->setIdv($idV);
         $this->setNomVoiture($nomVoiture);
         $this->setDescription($description);
@@ -30,7 +30,7 @@ class Voiture {
     function getImage():string {
         return $this->image;
     }
-    function getPrix():float {
+    function getPrix():double {
         return $this->prix;
     }
     function getQuantite():int {
